feat: mime types
To this feature, a map was created to store the correct content type to sento to the header. The key is the extension of the file.

Then, a search for the extension is made and a find in the map to get the correct Content-Type. If anything is found, the "text/pain" Content-Type is setted by default